style: 가위바위게임의 캡슐화, 은닉화 진행
